1. rename the moudle name from awsrest to awsrestv2 2. add v4_validate_reissue ability 3. TODO: session-token, change url, validation from query paramters
