+++ src/vmod_awsrestv2.vcc
@@ -1,0 +1,26 @@
+#
+# This document is licensed under the same license as the libvmod-awsrest project. See LICENSE for details.
+# Copyright (c) 2012-2017 Shohei Tanaka(@xcir)
+#
+# File layout and configuration based on libvmod-example
+# Copyright (c) 2011 Varnish Software AS
+# hmac-sha1 and base64 based on libvmod-digest( https://github.com/varnish/libvmod-digest )
+#
+
+$Module awsrestv2 Awsrest VMOD
+$ABI vrt
+
+DESCRIPTION
+===========
+
+Validate the AWS request
+
+* Change from https://github.com/xcir/libvmod-awsrest
+* https://github.com/sophiah/libvmod-awsrest
+* https://docs.aws.amazon.com/general/latest/gr/signature-version-4.html
+
+
+$Event event_function
+
+$Function BOOL v4_validate(STRING access_key, STRING secret_key)
+$Function STRING v4_validate_reissue(STRING org_access_key, STRING org_secret_key, STRING new_access_key, STRING new_secret_key, STRING new_session_token = "", STRING new_url = "")
